+++ apps/desktop/src/app/router.tsx
@@ -3,7 +3,8 @@ import { createBrowserRouter } from "react-router-dom";
 import AppLayout from "@/components/layout/AppLayout";
 
 import DashboardPage from "@/features/dashboard/DashboardPage";
-import MonitoringPage from "@/features/monitoring/MonitoringPage";
+import MonitoringPage from "@/features/monitoring/pages/MonitoringPage";
+import StoragePage from "@/features/storage/pages/StoragePage";
 
 export const router = createBrowserRouter([
   {
@@ -17,6 +18,10 @@ export const router = createBrowserRouter([
       {
         path: "monitoring",
         element: <MonitoringPage />,
+      },
+      {
+        path: "storage",
+        element: <StoragePage />,
       },
     ],
   },